+++ CHANGELOG.md
@@ -89,6 +89,7 @@ New deprecation(s):
 
 ### Other
 
+- **General**: Add e2e test for external push scaler ([#2698](https://github.com/kedacore/keda/pull/2698))
 - **General**: Bump Golang to 1.20 ([#4517](https://github.com/kedacore/keda/issues/4517))
 - **General**: Drop a transitive dependency on bou.ke/monkey ([#4364](https://github.com/kedacore/keda/issues/4364))
 - **General**: Fix odd number of arguments passed as key-value pairs for logging ([#4368](https://github.com/kedacore/keda/issues/4368))